pyproject.toml
[tool.poetry]
name = "goodreads-visualizer"
version = "0.1.0"
description = ""
authors = ["Jordan Katzen <jkatzen8@gmail.com>"]
readme = "README.md"
packages = [{include = "goodreads_visualizer"}]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.25.0"
beautifulsoup4 = "^4.12.2"
supabase = "^1.0.4"
python-dotenv = "^1.0.0"
Flask = "^3.0.0"
gunicorn = "^21.2.0"
requests = "^2.31.0"
pytz = "^2023.3.post1"
feedparser = "^6.0.11"

[tool.poetry.dev-dependencies]
ipython = "^8.14.0"
ipykernel = "^6.24.0"
ruff = "^0.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
